==> step9-label_choose/1.data.ipynb <==
import numpy as np
import pandas as pd
similarity_matrix = np.load('/data/chenxi/research/conference/TMIS/COTR/step3-bert_recall/CuBERT/outputs/en/2e-05_32_0.8/similarity_matrix.npy').T
job_df = pd.read_csv('/data/chenxi/research/conference/TMIS/COTR/step3-bert_recall/data/en/en.csv', sep='\t')
task_df = pd.read_csv('/data/chenxi/research/conference/TMIS/COTR/step3-bert_recall/data/en/task.csv', sep='\t')

task_map = pd.read_excel('/data/chenxi/research/conference/TMIS/COTR/step4-llm_select/results/en/en_results.xlsx')
task_map = task_map.groupby('task').agg(list)[['job_description']].reset_index()
task_df['task'] = task_df['DWA Title']
task_df = pd.merge(task_df, task_map, how='left').reset_index(drop=True)

index_list = np.argsort(-similarity_matrix)
sort_similarity_matrix = -np.sort(-similarity_matrix)
top_num = 20
task_df['sort_job'] = [i[:top_num] for i in index_list]
task_df['sort_job'] = task_df['sort_job'].apply(lambda x: [job_df['job_description'][i] for i in x])
task_df['sort_sim'] = [i[:top_num] for i in sort_similarity_matrix]
task_df['job_len'] = task_df['job_description'].apply(lambda x: len(x) if type(x) == type([]) else 0)
task_df

save_df = task_df[~task_df['job_description'].isna()].drop('task', axis=1).sort_values(['Element Name', 'IWA Title', 'DWA Title', 'job_len'], ascending=[False, False, False, False])
save_df.to_excel('data/en/task_jd.xlsx', index=None)

num_clusters = 100

from sklearn.metrics.pairwise import cosine_distances

def compute_distance_matrix(matrix):
    # 计算余弦距离
    distance_matrix = cosine_distances(matrix)
    return distance_matrix

from scipy.cluster.hierarchy import linkage, fcluster

def hierarchical_clustering(distance_matrix, num_clusters=35):
    # 使用层次聚类算法
    Z = linkage(distance_matrix, method='ward')  # 'ward' 方法适用于最小化欧式距离平方
    # 将标签分成num_clusters个簇
    clusters = fcluster(Z, t=num_clusters, criterion='maxclust')
    return clusters

import numpy as np

def select_representative_labels(distance_matrix, clusters, num_labels=35):
    unique_clusters = np.unique(clusters)
    representative_labels = []
    
    for cluster_id in unique_clusters:
        cluster_indices = np.where(clusters == cluster_id)[0]
        avg_distances = np.mean(distance_matrix[cluster_indices][:, cluster_indices], axis=1)
        representative_label = cluster_indices[np.argmax(avg_distances)]
        representative_labels.append(representative_label)

    return representative_labels[:num_labels]

distance_matrix = compute_distance_matrix(similarity_matrix)

clusters = hierarchical_clustering(distance_matrix, num_clusters=num_clusters)

representative_labels = select_representative_labels(distance_matrix, clusters, num_labels=num_clusters)

print("Representative labels:", representative_labels)

choose_task = task_df.loc[representative_labels]
choose_task = choose_task.sort_values(['job_len', 'Element Name', 'IWA Title'], ascending=[False, False, False]).reset_index(drop=True)
choose_task = choose_task.drop_duplicates('IWA Title')
choose_task = choose_task.explode('job_description')[['Element Name', 'IWA Title', 'DWA Title', 'job_description']]
choose_task.to_excel('data/en/choose_task_v1.xlsx', index=None)
choose_task

large_task_list = task_df[task_df['job_len']>20].index.tolist()
choose_task2 = task_df.loc[list(set(large_task_list) - set(representative_labels))]
choose_task2 = choose_task2.sort_values([ 'job_len', 'Element Name', 'IWA Title', 'DWA Title'], ascending=[False, False, False, False]).reset_index(drop=True)
choose_task2 = choose_task2.explode('sort_job')[['Element Name', 'IWA Title', 'DWA Title', 'sort_job']]
choose_task2.to_excel('data/en/choose_task_v2.xlsx', index=None)
choose_task2

choose_task_list = ["Contact current or potential customers to promote products or services.","Evaluate applicable laws and regulations to determine impact on organizational activities.","Confer with managers to make operational decisions.","Communicate with customers to resolve complaints or ensure satisfaction.","Develop promotional strategies or plans."]
choose_task3 = task_df[task_df['task'].isin(choose_task_list)]
choose_task3 = choose_task3.sort_values([ 'job_len', 'Element Name', 'IWA Title', 'DWA Title'], ascending=[False, False, False, False]).reset_index(drop=True)
choose_task3_1 = choose_task3.explode('job_description')[['Element Name', 'IWA Title', 'DWA Title', 'job_description']]
choose_task3_2 = choose_task3.explode('sort_job')[['Element Name', 'IWA Title', 'DWA Title', 'sort_job']]
choose_task3_2.columns = choose_task3_1.columns
choose_task3 = pd.concat([choose_task3_1, choose_task3_2])
choose_task3.sort_values(['Element Name', 'IWA Title', 'DWA Title']).drop_duplicates()
choose_task3.to_excel('data/en/choose_task_v3.xlsx', index=None)
choose_task3

large_task_list = task_df[(task_df['job_len']>10) & (task_df['job_len']<=20)].index.tolist()
choose_task2 = task_df.loc[list(set(large_task_list) - set(representative_labels))]
choose_task2 = choose_task2.sort_values([ 'job_len', 'Element Name', 'IWA Title', 'DWA Title'], ascending=[False, False, False, False]).reset_index(drop=True)
choose_task2 = choose_task2.explode('sort_job')[['Element Name', 'IWA Title', 'DWA Title', 'sort_job']]
choose_task2.to_excel('data/en/choose_task_v4.xlsx', index=None)
choose_task2

import pandas as pd
pd.read_excel('')
