==> step9-label_choose/2.human.ipynb <==
import pandas as pd
dwa_ch = pd.read_excel('data/en/DWA Reference-ch.xlsx')
dwa = pd.read_excel('data/en/DWA Reference.xlsx')

dwa

dwa_ch

map_ch = {i:v for i,v in zip(dwa_ch['DWA 标题'], dwa['DWA Title'])}
label_list_ch = [i.strip('\n') for i in open('data/en/step1/label_ch.txt').readlines()]
label_list = [map_ch[i] for i in label_list_ch]

pd.DataFrame([label_list,label_list_ch]).T[0].tolist()
