==> step9-label_choose/3.filter.ipynb <==
import pandas as pd
df = pd.read_excel('../step1-job_task_extraction/results/en/en-job_re_description.xlsx')
origin_df = pd.read_csv('../data/en/job_description.csv', sep='\t')
# df.to_csv('/data/chenxi/research/conference/TMIS/COTR/step1-job_task_extraction/results/en/en-job_re_description.csv', index=None)

task_list = ['Contact current or potential customers to promote products or services.',
 'Evaluate applicable laws and regulations to determine impact on organizational activities.',
 'Develop promotional strategies or plans.',
 'Manage budgets for appropriate resource allocation.',
 'Train patients, family members, or caregivers in techniques for managing disabilities or illnesses.',
 'Prepare technical reports for internal use.',
 'Inspect facilities to ensure compliance with safety, quality, or service standards.',
 'Develop instructional materials.',
 'Monitor operational activities to ensure compliance with regulations or standard operating procedures.',
 'Represent the organization in external relations.']

rules = [r"" for i in range(10)] 
rules[3] = r"Manage.*?budgets"
rules[4] = r"((train|Support|education).*?Patient care)|(Patient care.*?(Management|education|Techniques))"
rules[5] = r"Technical Writing|Research and Development|Documentation Standards|Technical Presentations"
rules[6] = r"Conduct Inspections|Identify Hazards|Identify Compliance Issues|Facility Management|Quality Assurance"
rules[7] = r"Content Creation|Content Development|Courseware|Training Manuals|Educational Resources|Curriculum Development|Instructional Design|Training Material Creation|Educational Content|Learning Material"
rules[8] = r"compliance monitoring|Ensuring Regulatory Compliance|Internal Auditing|Compliance Audits|Regulatory Review|Regulatory"
rules[9] = r"maintaining relationships|representing the|representive|interfacing"

import re
all_df = []
for i in range(3, 10):
    df[task_list[i]] = df['task'].apply(lambda x: len(re.findall(rules[i], x, re.IGNORECASE)) > 0)
    sub_df = df[df[task_list[i]]][['job_title', 'task']]
    sub_df = sub_df.sample(frac=1.0)
    sub_df['job_description'] = sub_df['task']
    sub_df.to_excel(f'outputs/key/{task_list[i]}.xlsx')
    sub_df['task'] = task_list[i]
    all_df.append(sub_df)
all_df = pd.concat(all_df)[['task', 'job_title', 'job_description']].reset_index(drop=True)
all_df.to_excel('outputs/key_extraction.xlsx', index=None)
all_df 

all_df_ch = pd.read_excel('outputs/key_extraction-ch.xlsx')
final_df = pd.concat([all_df, all_df_ch], axis=1)
final_df.to_excel('outputs/key_extraction-merge.xlsx', index=None)
final_df

import re
all_df = []
for i in range(3, 10):
    origin_df[task_list[i]] = origin_df['job_description'].apply(lambda x: len(re.findall(rules[i], x, re.IGNORECASE)) > 0)
    sub_df = origin_df[origin_df[task_list[i]]][['job_title', 'job_description']]
    sub_df = sub_df.sample(frac=1.0)
    sub_df['task'] = task_list[i]
    all_df.append(sub_df)
all_df = pd.concat(all_df)[['task', 'job_title', 'job_description']].reset_index(drop=True)
all_df.to_excel('outputs/key_extraction_fromjd.xlsx', index=None)
all_df 

from tqdm import tqdm
def split_sentence_by_uppercase(sentence):
    # 在大写字母前插入空格
    return re.sub(r'([A-Z])', r' \1', sentence).strip()

tqdm.pandas()
all_df = pd.read_excel('outputs/key_extraction_fromjd.xlsx')
all_df['extract_task'] = all_df.progress_apply(lambda x: [line for line in re.findall(r'[A-Z][^A-Z]*', split_sentence_by_uppercase(x['job_description'])) if re.search(rules[task_list.index(x['task'])], line, re.IGNORECASE) ] if type(x['job_description']) == type('') else [], axis=1)
all_df.to_excel('outputs/key_extraction_fromjd_annotated.xlsx', index=None)
