==> step9-label_choose/4.filter.ipynb <==
import pandas as pd
dataset_name = 'en'
step_turn = 'step2'
seg_df = pd.read_excel('../step1-job_task_extraction/results/en/en-job_re_description.xlsx')
origin_df = pd.read_csv('../data/en/job_description.csv', sep='\t')
df_map = {'jd_seg':seg_df, 'jd_ori': origin_df}

import json
task_kind = 'task'
task_map = json.load(open(f'data/en/{step_turn}/{task_kind}.json', 'r'))
task_rule_map = {i:'|'.join(v) for i,v in task_map.items()}
task_rule_map = {i:f"{v}" for i,v in task_rule_map.items()}
task_rule_map

df_map = {'jd_seg':seg_df, 'jd_ori': origin_df}
import re
import os
import copy
from tqdm import tqdm
tqdm.pandas()
cal_num = {}
def split_sentence_by_uppercase(sentence):
    # 在大写字母前插入空格
    return re.sub(r'([A-Z])', r' \1', sentence).strip()

if not os.path.exists(f"outputs/{step_turn}"):
    os.makedirs(f"outputs/{step_turn}")
for df_type, df_bp in df_map.items():
    all_df = []
    cal_num[df_type] = {}
    if os.path.exists(f'data/en/{step_turn}/{df_type}.parquet'):
        df = pd.read_parquet(f'data/en/{step_turn}/{df_type}.parquet')
    else:
        df = copy.deepcopy(df_bp)
        if df_type == 'jd_seg':
            df['job_description'] = df['task']
        df['job_description'] = df.apply(lambda x: [line for line in re.findall(r'[A-Z][^A-Z]*', split_sentence_by_uppercase(x['job_description']))] if type(x['job_description']) == type('') else [], axis=1)
        df = df.explode('job_description')
        df = df[~df['job_description'].isna()]
        df['len'] = df['job_description'].apply(lambda x: len(x.split(' ')))
        df = df[(df['len']>6) & (df['len']<20)]
        df = df.drop_duplicates('job_description')
        df.to_parquet(f'data/en/{step_turn}/{df_type}.parquet', index=None)

    for task, task_rule in tqdm(task_rule_map.items(), total=len(task_rule_map)):
        if len(task_rule) == 0:
            continue            
        df[task] = df['job_description'].apply(lambda x: len(re.findall(task_rule, x, re.IGNORECASE)) > 0)
        sub_df = df[df[task]][['job_title', 'job_description']]
        print(df_type, task, len(sub_df))
        cal_num[df_type][task] = len(sub_df)
        # sub_df = sub_df.sample(frac=1.0)
        sub_df['task'] = task
        all_df.append(sub_df)
    all_df = pd.concat(all_df)[['task', 'job_title', 'job_description']].reset_index(drop=True)
    all_df.to_excel(f'outputs/{step_turn}/key_extraction_{df_type}-{task_kind}.xlsx', index=None)
cal_num = pd.DataFrame(cal_num)
cal_num

cal_num = {}
for df_type, df_bp in df_map.items():
    df = pd.read_excel(f'outputs/{step_turn}/key_extraction_{df_type}.xlsx')
    cal_num[df_type] = {}
    for task, task_rule in task_rule_map.items():
        if len(task_rule) == 0:
            continue   
        cal_num[df_type][task] = len(df[df['task']==task])
cal_num = pd.DataFrame(cal_num)
cal_num

import pandas as pd
cal_num = {}
final_df = None
for df_type, df_bp in df_map.items():
    df = pd.read_excel(f'outputs/{step_turn}/key_extraction_{df_type}.xlsx')
    if final_df is None:
        final_df = df
    else:
        final_df_group = final_df.groupby('task').agg(list)
        final_df_group['len'] = final_df_group['job_description'].apply(len)
        df_group = df.groupby('task').agg(list)
        df_group['len'] = df_group['job_description'].apply(len)
        final_df = pd.concat(final_df)
cal_num = pd.DataFrame(cal_num)
cal_num

# tqdm.pandas()
# all_df = pd.read_excel('outputs/key_extraction_fromjd.xlsx')
# all_df['extract_task'] = all_df.progress_apply(lambda x: [line for line in re.findall(r'[A-Z][^A-Z]*', split_sentence_by_uppercase(x['job_description'])) if re.search(rules[task_list.index(x['task'])], line, re.IGNORECASE)] if type(x['job_description']) == type('') else [], axis=1)
# all_df.to_excel('outputs/key_extraction_fromjd_annotated.xlsx', index=None)
# all_df_ch = pd.read_excel('outputs/key_extraction-ch.xlsx')
# final_df = pd.concat([all_df, all_df_ch], axis=1)
# final_df.to_excel('outputs/key_extraction-merge.xlsx', index=None)
# final_df
