==> step9-label_choose/5.data.ipynb <==
import pandas as pd
step1_seg_seg = pd.read_excel('/data/chenxi/research/conference/TMIS/COTR/step9-label_choose/outputs/step1/key_extraction.xlsx')
step1_ori_seg = pd.read_excel('/data/chenxi/research/conference/TMIS/COTR/step9-label_choose/outputs/step1/key_extraction_fromjd.xlsx')
step2_ori_jd = pd.read_excel('/data/chenxi/research/conference/TMIS/COTR/step9-label_choose/outputs/step2/key_extraction_jd_ori.xlsx')
step2_seg_jd = pd.read_excel('/data/chenxi/research/conference/TMIS/COTR/step9-label_choose/outputs/step2/key_extraction_jd_seg.xlsx')

df = pd.concat([step1_seg_seg, step1_ori_seg, step2_ori_jd, step2_seg_jd])
df = df[~df['job_description'].isna()]
df = df.sample(3478)

df['job_title'].nunique()

rule = "|".join(df['job_description'].tolist())
rule = rf"{rule}"

origin_df = pd.read_csv('/data/chenxi/research/conference/TMIS/COTR/data/en/job_description.csv', sep='\t')
origin_df

from tqdm import tqdm
import re
tqdm.pandas()
jd_list = df['job_description'].tolist()
sub_df = origin_df[:]
sub_df['isin'] = sub_df['job_description'].progress_apply(lambda x: [y for y in jd_list if y in x])
sub_df = sub_df[sub_df['isin'].apply(len)!=0]
print(len(sub_df))

sub_df

sub_df['job_title'].nunique()
